requirements.txt
google-api-python-client==1.1
httplib2==0.8
keyring==1.2.2
python-gflags==2.0
simplejson==3.1.3
wsgiref==0.1.2
